Hack-ish way of fixing scriptengine for now
Fixed up hot reloading
Replaced MSBuild.exe hardcoded pathing using vswhere.exe
