+++ binfiles/Assets/NewSceneEmpty.scene
@@ -1,0 +1,3 @@
+{
+    "Entities": []
+}
